requirements.txt
Flask==1.1.2
requests==2.25.1
Flask_SQLAlchemy==2.4.4
Flask_Bcrypt==0.7.1
termcolor==1.1.0
PyJWT==2.1.0
